microservices_python/requirements.txt
beautifulsoup4
requests
JayDeBeApi
pymongo
flask
flask_cors
schedule
python-dateutil
JPype1==0.6.3